+++ core/src/main/java/fr/abes/licencesnationales/core/repository/EtablissementRepository.java
@@ -12,7 +12,7 @@ import java.util.Optional;
 
 
 @Repository
-public interface EtablissementRepository extends JpaRepository<EtablissementEntity, Long> , JpaSpecificationExecutor<EtablissementEntity> {
+public interface EtablissementRepository extends JpaRepository<EtablissementEntity, Integer> , JpaSpecificationExecutor<EtablissementEntity> {
 
     void deleteBySiren(String siren);
