+++ core/src/main/java/fr/abes/licencesnationales/core/repository/ip/IpV6Repository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IpV6Repository extends JpaRepository<IpV6, Long> {
+public interface IpV6Repository extends JpaRepository<IpV6, Integer> {
 
 }
